fix: use mock test for preset
